+++ examples/src/examples/thingie.clj
@@ -1,7 +1,7 @@
 (ns examples.thingie
   (:require [ring.util.http-response :refer :all]
             [compojure.api.sweet :refer :all]
-            [compojure.api.upload :refer :all]
+            [ring.swagger.upload :refer :all]
             [schema.core :as s]
             ring.swagger.json-schema-dirty
             [examples.pizza :refer [pizza-routes Pizza]]
@@ -99,7 +99,7 @@
     :tags ["responses"]
     (GET* "/" []
       :query-params [return :- (s/enum :200 :403 :404)]
-      :responses    {403 {:schema {:code s/Str, :description "spiders?"}}
+      :responses    {403 {:schema {:code s/Str}, :description "spiders?"}
                      404 {:schema {:reson s/Str}, :description "lost?"}}
       :return       Total
       :summary      "multiple returns models"