+++ CHANGELOG.md
@@ -137,7 +137,7 @@
 
 ## 0.20.0 (24.4.2015)
 
-* New restructuring for `:no-docs` (a boolean) - endpoints with this don't get api documentation.
+* New restructuring for `:no-doc` (a boolean) - endpoints with this don't get api documentation.
 * Fixed [#42](https://github.com/metosin/compojure-api/issues/42) - `defroutes*` now does namespace resolution for the source
 used for route peeling and source linking (the macro magic)
 * Fixed [#91](https://github.com/metosin/compojure-api/issues/91) - `defroutes*` are now automatically accessed over a Var for better development flow.