+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ```clojure
 [cheshire "5.5.0"] is available but we use "5.4.0"
+[backtick "0.3.3"] is available but we use "0.3.2"
 ```
 
 ## 0.21.0 (25.5.2015)